Fix compiler warning C4267: 'initializing' : conversion from 'size_t' to 'int', possible loss of data Fix compiler warning C4267: '=' : conversion from 'size_t' to 'int', possible loss of data
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7191 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/AlbumFolder.cpp
@@ -858,7 +858,7 @@ BOOL CAlbumFolder::OrganiseFile(CLibraryFile* pFile)
 				split_results splitResults;
 				std::wstring sFileName( pFile->m_sName );
 
-				int nCount = firstPattern.split( sFileName, splitResults, 0 );
+				size_t nCount = firstPattern.split( sFileName, splitResults, 0 );
 				if ( nCount < 4 )
 				{
 					nCount = secondPattern.split( sFileName, splitResults, 0 );
